+++ src/core/jvm/instructions/CastInstruction.js
@@ -1,21 +1,8 @@
-const AbstractInstruction = require('./AbstractInstruction');
+const InstructionWrapper = require('./InstructionWrapper');
 
-export default class CastInstruction extends AbstractInstruction {
+export default class CastInstruction extends InstructionWrapper {
 
   constructor(instruction) {
-    super(instruction.idx, instruction.opcode);
-    this.instruction = instruction;
-  }
-
-  get size() {
-    return this.instruction.size;
-  }
-
-  read(buffer) {
-    this.instruction.read(buffer);
-  }
-
-  write(buffer) {
-    this.instruction.write(buffer);
+    super(instruction);
   }
 }
